Fix a lockout
Remove a clear watchdog in the second calibrate loop to prevent a
lockout when the calibration is out of boundaries and the exit condition
is never met.
The watchdog will give 256ms for the loop to exit (which is more than
enough if everything goes well), otherwise it will reset and re-do the
calibration again.