Ball detection 30sec timeout
Timeout to prevent lock-on switches.
The switch will force a re-calibration and go back to the off-state.
The downside is that this inductive switch cannot be used on a game
where a ball can sit on the switch for more than 30s (it is not the case
on Full Throttle).
The live calibration will ensure the switch will work correctly
afterwards.